Changes to Universal Products

+++ Plugins/RFMap/RFMap.m
@@ -68,7 +68,7 @@ LLTaskPlugIn	*task = nil;
 
 @implementation RFMap
 
-+ (long)version;
++ (NSInteger)version;
 {
 	return kLLPluginVersion;
 }
